# ExplorerFrontend/app/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZondScan Public API",
    "version": "1.1.0",
...
  },
  "servers": [{ "url": "https://zondscan.com" }],
...
  "paths": {
...
    "/api/contract/verify": {
      "post": {
        "tags": ["Contract Verification"],
        "operationId": "verifyContract",
        "summary": "Submit contract source for verification",
        "description": "Enqueues an async source-verification job compiled with the Hyperion toolchain. Request bodies are capped at 1 MiB. Rate limited per IP: bucket of 5 requests, refilling 5 per minute. Poll GET /api/contract/verify/{jobId} for the outcome.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": { "type": "string", "description": "Canonical QIP-55 contract address (uppercase Q plus 128 checksum-cased hex characters)." },
                  "sourceCode": { "type": "string" },
                  "contractName": { "type": "string", "pattern": "^[A-Za-z_$][A-Za-z0-9_$]*$", "description": "Hyperion identifier. Dollar signs are valid in both the first and subsequent positions." },
                  "compilerVersion": { "type": "string", "description": "Build id from /api/contract/compiler-info. Empty selects the default build." },
                  "optimizerEnabled": { "type": "boolean" },
                  "optimizerRuns": { "type": "integer", "description": "Effective optimizer run count. When optimization is enabled, zero or a negative value is canonicalized to 200 before enqueue, compilation, and persistence." },
                  "evmVersion": { "type": "string" },
                  "constructorArguments": { "type": "string" },
                  "libraries": { "type": "object", "additionalProperties": { "type": "string" } },
                  "imports": {
                    "type": "object",
                    "description": "Additional Hyperion sources keyed by clean relative forward-slash paths. Paths are canonicalized before compilation and persistence; aliases, absolute paths, root escapes, and collisions with <contractName>.hyp are rejected.",
                    "additionalProperties": { "type": "string" }
                  },
                  "license": { "type": "string" }
                },
                "required": ["address", "sourceCode", "contractName"]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": { "type": "string" },
                    "status": { "type": "string" },
                    "address": { "type": "string" }
                  }
                },
                "example": { "jobId": "9f2c4a6e8b0d1f3a", "status": "pending", "address": "QCCCccccCcCcCccCcCccCCCccCcCcCcccccCCccCccCccccCcCCcCcccCCcCcccccccCccCcCccccCCcCccCccCcccCcCCCccCcCccCcCccccccCcCcCcccCcCcCCCccC" }
              }
            }
          },
          "200": {
            "description": "The contract is already verified; no new job was created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "alreadyVerified": { "type": "boolean" },
                    "address": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, invalid address, missing required fields, or unsupported compilerVersion (the body lists supported builds).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "409": {
            "description": "A verification job is already in flight, or the stored verification record is invalid and requires operator repair. Public re-verification never overwrites an invalid stored record.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": {
            "description": "Contract verification is not configured on this deployment.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/contract/verify/{jobId}": {
      "get": {
        "tags": ["Contract Verification"],
        "operationId": "getVerificationJob",
        "summary": "Verification job status",
        "description": "Returns the verification job document. Status moves through pending, compiling, and then success or failed.",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "description": "Job id returned by POST /api/contract/verify.", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Job document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": { "type": "string" },
                    "address": { "type": "string" },
                    "status": { "type": "string", "enum": ["pending", "compiling", "success", "failed"] },
                    "error": { "type": "string" },
                    "createdAt": { "type": "string" },
                    "updatedAt": { "type": "string" },
                    "target": {
                      "type": "object",
                      "description": "Canonical deployment generation captured before compilation and revalidated before publication.",
                      "properties": {
                        "address": { "type": "string", "pattern": "^Q[0-9a-f]{128}$" },
                        "creationTransaction": { "type": "string", "pattern": "^(?:|0x[0-9a-f]{64})$" },
                        "creationBlockNumber": { "type": "string", "pattern": "^0x(?:0|[1-9a-f][0-9a-f]*)$" },
                        "creationBlockHash": { "type": "string", "pattern": "^0x[0-9a-f]{64}$" },
                        "chainId": { "type": "string", "pattern": "^0x(?:0|[1-9a-f][0-9a-f]*)$" },
                        "deployedCodeSha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
                        "genesisContract": { "type": "boolean" }
                      },
                      "required": ["address", "creationBlockNumber", "creationBlockHash", "chainId", "deployedCodeSha256", "genesisContract"],
                      "additionalProperties": false
                    },
                    "payload": {
                      "type": "object",
                      "description": "Echo of the submitted request.",
                      "properties": {
                        "sourceCode": { "type": "string" },
                        "contractName": { "type": "string" },
                        "compilerVersion": { "type": "string" },
                        "compilerProvenance": { "$ref": "#/components/schemas/CompilerProvenance" },
                        "optimizationEnabled": { "type": "boolean" },
                        "optimizationRuns": { "type": "integer" },
                        "evmVersion": { "type": "string" },
                        "constructorArguments": { "type": "string" },
                        "libraries": { "type": "object", "additionalProperties": { "type": "string" } },
                        "imports": {
                          "type": "object",
                          "description": "Canonical effective import sources used for compilation.",
                          "additionalProperties": { "type": "string" }
                        },
                        "license": { "type": "string" },
                        "verificationMethod": { "type": "string", "const": "full-source" }
                      },
                      "additionalProperties": true
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "abi": { "type": "string" },
                        "bytecodeHash": { "type": "string" },
                        "artifactDigest": {
                          "type": "string",
                          "pattern": "^qrl\\.contract-verification-artifact\\.v2:sha256:[0-9a-f]{64}$",
                          "description": "V2 digest binding the source bundle, ABI, compiler identity and settings, and canonical deployment target."
                        },
                        "compilerProvenance": { "$ref": "#/components/schemas/CompilerProvenance" },
                        "verifiedAt": { "type": "string" }
                      }
                    }
                  },
                  "required": ["jobId", "address", "status", "target", "payload"]
                }
              }
            }
          },
          "404": {
            "description": "Job not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/contract/compiler-info": {
      "get": {
        "tags": ["Contract Verification"],
        "operationId": "getCompilerInfo",
        "summary": "Available compiler builds",
        "description": "Lists every selectable Hyperion compiler build and identifies the default. Use a listed buildId as the compilerVersion in verify submissions.",
        "responses": {
          "200": {
            "description": "Compiler registry.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "language": { "type": "string" },
                    "buildId": { "type": "string" },
                    "kind": { "type": "string", "enum": ["native"] },
                    "sha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
                    "provenance": { "$ref": "#/components/schemas/CompilerProvenance" },
                    "default": { "type": "string" },
                    "compilers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "buildId": { "type": "string" },
                          "language": { "type": "string" },
                          "kind": { "type": "string", "enum": ["native"] },
                          "sha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
                          "provenance": { "$ref": "#/components/schemas/CompilerProvenance" },
                          "default": { "type": "boolean" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Contract verification is not configured on this deployment.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "InternalError": {
        "description": "Internal server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Per-IP rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "rate limit exceeded" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "CompilerProvenanceComponent": {
        "type": "object",
        "description": "One content-addressed runtime component that affected compiler output.",
        "required": ["name", "sha256"],
        "properties": {
          "name": { "type": "string", "enum": ["hypc", "nsjail", "policy"] },
          "sha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" }
        },
        "additionalProperties": false
      },
      "CompilerProvenance": {
        "type": "object",
        "description": "Content-addressed identity of the sealed native compiler, static NsJail launcher, and embedded sandbox policy persisted at verification time. Schema v2 binds the exact ordered components with a length-prefixed SHA-256 execution digest. Its absence on an older record means legacy-unrecorded and must not be inferred from the current registry.",
        "required": ["schema", "kind", "buildId", "executionDigest", "components"],
        "properties": {
          "schema": { "type": "string", "const": "qrl.contract-compiler-provenance.v2" },
          "kind": { "type": "string", "const": "native" },
          "buildId": { "type": "string", "minLength": 1 },
          "executionDigest": { "type": "string", "pattern": "^[0-9a-f]{64}$" },
          "components": {
            "type": "array",
            "minItems": 3,
            "maxItems": 3,
            "prefixItems": [
              {
                "allOf": [
                  { "$ref": "#/components/schemas/CompilerProvenanceComponent" },
                  { "type": "object", "properties": { "name": { "const": "hypc" } } }
                ]
              },
              {
                "allOf": [
                  { "$ref": "#/components/schemas/CompilerProvenanceComponent" },
                  { "type": "object", "properties": { "name": { "const": "nsjail" } } }
                ]
              },
              {
                "allOf": [
                  { "$ref": "#/components/schemas/CompilerProvenanceComponent" },
                  { "type": "object", "properties": { "name": { "const": "policy" } } }
                ]
              }
            ],
            "items": false
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}